Refactored habit tests and added aging during sleep
